requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow>=9.0.0
reportlab==4.0.4
openpyxl==3.1.2
django-extensions==3.2.3
python-decouple==3.8